+++ src/main/java/cart/controller/CartController.java
@@ -2,7 +2,8 @@ package cart.controller;
 
 import cart.auth.AuthAccount;
 import cart.global.annotation.LogIn;
-import cart.service.CartService;
+import cart.service.CartQueryService;
+import cart.service.CartCommandService;
 import cart.service.dto.ProductSearchResponse;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
@@ -18,10 +19,12 @@ import java.util.List;
 @Controller
 public class CartController {
 
-    private final CartService cartService;
+    private final CartCommandService cartCommandService;
+    private final CartQueryService cartQueryService;
 
-    public CartController(final CartService cartService) {
-        this.cartService = cartService;
+    public CartController(final CartCommandService cartCommandService, final CartQueryService cartQueryService) {
+        this.cartCommandService = cartCommandService;
+        this.cartQueryService = cartQueryService;
     }
 
     @PostMapping("/carts/products/{product-id}")
@@ -29,7 +32,7 @@ public class CartController {
     public String registerProductInCart(@LogIn AuthAccount account,
                                         @PathVariable("product-id") final Long productId) {
 
-        cartService.registerProductInCart(account, productId);
+        cartCommandService.registerProductInCart(account, productId);
 
         return "redirect:/";
     }
@@ -37,14 +40,14 @@ public class CartController {
     @GetMapping("/carts")
     @ResponseBody
     public List<ProductSearchResponse> showAllProductsInCart(@LogIn AuthAccount account) {
-        return cartService.findAllProductsInCart(account);
+        return cartQueryService.findAllProductsInCart(account);
     }
 
     @DeleteMapping("/carts/products/{product-id}")
     public String deleteProductInCart(@LogIn AuthAccount account,
                                       @PathVariable("product-id") final Long productId) {
 
-        cartService.deleteProductInCart(account, productId);
+        cartCommandService.deleteProductInCart(account, productId);
 
         return "cart";
     }
